refactored view and component file structure
