--- .husky/pre-push
@@ -1,4 +1,0 @@
-#!/bin/sh
-. "$(dirname "$0")/_/husky.sh"
-
-cd server && npm run checks && cd ../client && npm run checks && cd ..
